Test demonstrating block_type column width issue
Since most implementations of SQLite don't actually enforce the column width
written in their schema, this test case doesn't actually fail when saving the
value into the database. But if you ran it in Postgres, it would, because it's
trying to save a string (the encrypted value) of 110 characters into a column
with a max width of 78.

In order to demonstrate the issue in the SQLite test environment, then, we have
to introspect the database and compare the stated column width to the size of
what is saved.